+++ src/main/java/me/pajic/toolpouch/tooltip/PreviewExtensionPoint.java
@@ -1,9 +1,8 @@
 package me.pajic.toolpouch.tooltip;
-/*
 
 //? if neoforge {
 
-import com.misterpemodder.shulkerboxtooltip.api.neoforge.ShulkerBoxTooltipPlugin;
+/*import com.misterpemodder.shulkerboxtooltip.api.neoforge.ShulkerBoxTooltipPlugin;
 import net.neoforged.fml.ModLoadingContext;
 
 public class PreviewExtensionPoint {
@@ -15,5 +14,4 @@ public class PreviewExtensionPoint {
 		);
 	}
 }
-//?}
-*/
+*///?}